+++ lib/frontman/app.rb
@@ -165,11 +165,5 @@ module Frontman
       local_data[key] unless local_data.nil?
     end
 
-    class ExistingResourceError < StandardError
-      def self.create(url, resource)
-        new("Unable to redirect for #{url},
-             the resource #{resource.file_path} already exists on this URL")
-      end
-    end
   end
 end
